Remove old IARC implementation
