+++ lib/iarc/__init__.py
@@ -1,2 +1,1 @@
 import client  # noqa
-import utils  # noqa
