--- lib/iarc/templates/get_app_info.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<WEBSERVICE SERVICE_NAME="GET_APP_INFO" TYPE="REQUEST">
-  <ROW>
-    <FIELD NAME="password" VALUE="{{ password }}" TYPE="String" />
-    <FIELD NAME="submission_id" VALUE="{{ submission_id }}" TYPE="String" />
-    <FIELD NAME="security_code" VALUE="{{ security_code }}" TYPE="String" />
-    <FIELD NAME="company" VALUE="{{ company }}" TYPE="String" />
-    <FIELD NAME="platform" VALUE="{{ platform }}" TYPE="String" />
-  </ROW>
-</WEBSERVICE>
