--- lib/iarc/templates/set_storefront_data.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<WEBSERVICE SERVICE_NAME="SET_STOREFRONT_DATA" TYPE="REQUEST">
-  <ROW>
-    <FIELD NAME="AppUrl" VALUE="{{ app_url }}" TYPE="String" />
-    <FIELD NAME="password" VALUE="{{ password }}" TYPE="String" />
-    <FIELD NAME="storefront_company" VALUE="{{ company }}" TYPE="String" />
-    <FIELD NAME="platform" VALUE="{{ platform }}" TYPE="String" />
-    <FIELD NAME="submission_id" VALUE="{{ submission_id }}" TYPE="Int" />
-    <FIELD NAME="security_code" VALUE="{{ security_code }}" TYPE="String" />
-    <FIELD NAME="rating_system" VALUE="{{ rating_system }}" TYPE="String" />
-    <FIELD NAME="release_date" VALUE="{{ release_date and release_date.strftime('%Y-%m-%d') or '' }}" TYPE="String" />
-    <FIELD NAME="storefront_title" VALUE="{{ title }}" TYPE="String" />
-    <FIELD NAME="storefront_rating" VALUE="{{ rating }}" TYPE="String" />
-    <FIELD NAME="storefront_descriptors" VALUE="{{ descriptors }}" TYPE="String" />
-    <FIELD NAME="storefront_interactive_elements" VALUE="{{ interactive_elements }}" TYPE="String" />
-    <FIELD NAME="notes" VALUE="" TYPE="String" />
-  </ROW>
-</WEBSERVICE>
