--- lib/iarc/wsdl/prod/iarc_services.wsdl
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><wsdl:definitions name="IARCServices" targetNamespace="http://tempuri.org/" xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/" xmlns:wsx="http://schemas.xmlsoap.org/ws/2004/09/mex" xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd" xmlns:wsa10="http://www.w3.org/2005/08/addressing" xmlns:wsp="http://schemas.xmlsoap.org/ws/2004/09/policy" xmlns:wsap="http://schemas.xmlsoap.org/ws/2004/08/addressing/policy" xmlns:msc="http://schemas.microsoft.com/ws/2005/12/wsdl/contract" xmlns:soap12="http://schemas.xmlsoap.org/wsdl/soap12/" xmlns:wsa="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:wsam="http://www.w3.org/2007/05/addressing/metadata" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:tns="http://tempuri.org/" xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/" xmlns:wsaw="http://www.w3.org/2006/05/addressing/wsdl" xmlns:soapenc="http://schemas.xmlsoap.org/soap/encoding/"><wsp:Policy wsu:Id="BasicHttpBinding_IIARCServices_policy"><wsp:ExactlyOne><wsp:All><sp:TransportBinding xmlns:sp="http://schemas.xmlsoap.org/ws/2005/07/securitypolicy"><wsp:Policy><sp:TransportToken><wsp:Policy><sp:HttpsToken RequireClientCertificate="false"/></wsp:Policy></sp:TransportToken><sp:AlgorithmSuite><wsp:Policy><sp:Basic256/></wsp:Policy></sp:AlgorithmSuite><sp:Layout><wsp:Policy><sp:Strict/></wsp:Policy></sp:Layout></wsp:Policy></sp:TransportBinding></wsp:All></wsp:ExactlyOne></wsp:Policy><wsdl:types><xsd:schema targetNamespace="http://tempuri.org/Imports"><xsd:import schemaLocation="https://www.globalratings.com/IARCPRODService/IARCServices.svc?xsd=xsd0" namespace="http://tempuri.org/"/><xsd:import schemaLocation="https://www.globalratings.com/IARCPRODService/IARCServices.svc?xsd=xsd1" namespace="http://schemas.microsoft.com/2003/10/Serialization/"/><xsd:import schemaLocation="https://www.globalratings.com/IARCPRODService/IARCServices.svc?xsd=xsd2" namespace="http://schemas.microsoft.com/2003/10/Serialization/Arrays"/><xsd:import schemaLocation="https://www.globalratings.com/IARCPRODService/IARCServices.svc?xsd=xsd3" namespace="http://schemas.datacontract.org/2004/07/IARC.Library.DAL"/><xsd:import schemaLocation="https://www.globalratings.com/IARCPRODService/IARCServices.svc?xsd=xsd4" namespace="http://schemas.datacontract.org/2004/07/IARC.Library.Business.Common"/></xsd:schema></wsdl:types><wsdl:message name="IIARCServices_Get_App_Info_InputMessage"><wsdl:part name="parameters" element="tns:Get_App_Info"/></wsdl:message><wsdl:message name="IIARCServices_Get_App_Info_OutputMessage"><wsdl:part name="parameters" element="tns:Get_App_InfoResponse"/></wsdl:message><wsdl:message name="IIARCServices_Set_Storefront_Data_InputMessage"><wsdl:part name="parameters" element="tns:Set_Storefront_Data"/></wsdl:message><wsdl:message name="IIARCServices_Set_Storefront_Data_OutputMessage"><wsdl:part name="parameters" element="tns:Set_Storefront_DataResponse"/></wsdl:message><wsdl:message name="IIARCServices_Get_Rating_Changes_InputMessage"><wsdl:part name="parameters" element="tns:Get_Rating_Changes"/></wsdl:message><wsdl:message name="IIARCServices_Get_Rating_Changes_OutputMessage"><wsdl:part name="parameters" element="tns:Get_Rating_ChangesResponse"/></wsdl:message><wsdl:message name="IIARCServices_Set_Override_InputMessage"><wsdl:part name="parameters" element="tns:Set_Override"/></wsdl:message><wsdl:message name="IIARCServices_Set_Override_OutputMessage"><wsdl:part name="parameters" element="tns:Set_OverrideResponse"/></wsdl:message><wsdl:message name="IIARCServices_CheckValidCertficateForOverride_InputMessage"><wsdl:part name="parameters" element="tns:CheckValidCertficateForOverride"/></wsdl:message><wsdl:message name="IIARCServices_CheckValidCertficateForOverride_OutputMessage"><wsdl:part name="parameters" element="tns:CheckValidCertficateForOverrideResponse"/></wsdl:message><wsdl:portType name="IIARCServices"><wsdl:operation name="Get_App_Info"><wsdl:input wsaw:Action="http://tempuri.org/IIARCServices/Get_App_Info" message="tns:IIARCServices_Get_App_Info_InputMessage"/><wsdl:output wsaw:Action="http://tempuri.org/IIARCServices/Get_App_InfoResponse" message="tns:IIARCServices_Get_App_Info_OutputMessage"/></wsdl:operation><wsdl:operation name="Set_Storefront_Data"><wsdl:input wsaw:Action="http://tempuri.org/IIARCServices/Set_Storefront_Data" message="tns:IIARCServices_Set_Storefront_Data_InputMessage"/><wsdl:output wsaw:Action="http://tempuri.org/IIARCServices/Set_Storefront_DataResponse" message="tns:IIARCServices_Set_Storefront_Data_OutputMessage"/></wsdl:operation><wsdl:operation name="Get_Rating_Changes"><wsdl:input wsaw:Action="http://tempuri.org/IIARCServices/Get_Rating_Changes" message="tns:IIARCServices_Get_Rating_Changes_InputMessage"/><wsdl:output wsaw:Action="http://tempuri.org/IIARCServices/Get_Rating_ChangesResponse" message="tns:IIARCServices_Get_Rating_Changes_OutputMessage"/></wsdl:operation><wsdl:operation name="Set_Override"><wsdl:input wsaw:Action="http://tempuri.org/IIARCServices/Set_Override" message="tns:IIARCServices_Set_Override_InputMessage"/><wsdl:output wsaw:Action="http://tempuri.org/IIARCServices/Set_OverrideResponse" message="tns:IIARCServices_Set_Override_OutputMessage"/></wsdl:operation><wsdl:operation name="CheckValidCertficateForOverride"><wsdl:input wsaw:Action="http://tempuri.org/IIARCServices/CheckValidCertficateForOverride" message="tns:IIARCServices_CheckValidCertficateForOverride_InputMessage"/><wsdl:output wsaw:Action="http://tempuri.org/IIARCServices/CheckValidCertficateForOverrideResponse" message="tns:IIARCServices_CheckValidCertficateForOverride_OutputMessage"/></wsdl:operation></wsdl:portType><wsdl:binding name="BasicHttpBinding_IIARCServices" type="tns:IIARCServices"><wsp:PolicyReference URI="#BasicHttpBinding_IIARCServices_policy"/><soap:binding transport="http://schemas.xmlsoap.org/soap/http"/><wsdl:operation name="Get_App_Info"><soap:operation soapAction="http://tempuri.org/IIARCServices/Get_App_Info" style="document"/><wsdl:input><soap:body use="literal"/></wsdl:input><wsdl:output><soap:body use="literal"/></wsdl:output></wsdl:operation><wsdl:operation name="Set_Storefront_Data"><soap:operation soapAction="http://tempuri.org/IIARCServices/Set_Storefront_Data" style="document"/><wsdl:input><soap:body use="literal"/></wsdl:input><wsdl:output><soap:body use="literal"/></wsdl:output></wsdl:operation><wsdl:operation name="Get_Rating_Changes"><soap:operation soapAction="http://tempuri.org/IIARCServices/Get_Rating_Changes" style="document"/><wsdl:input><soap:body use="literal"/></wsdl:input><wsdl:output><soap:body use="literal"/></wsdl:output></wsdl:operation><wsdl:operation name="Set_Override"><soap:operation soapAction="http://tempuri.org/IIARCServices/Set_Override" style="document"/><wsdl:input><soap:body use="literal"/></wsdl:input><wsdl:output><soap:body use="literal"/></wsdl:output></wsdl:operation><wsdl:operation name="CheckValidCertficateForOverride"><soap:operation soapAction="http://tempuri.org/IIARCServices/CheckValidCertficateForOverride" style="document"/><wsdl:input><soap:body use="literal"/></wsdl:input><wsdl:output><soap:body use="literal"/></wsdl:output></wsdl:operation></wsdl:binding><wsdl:service name="IARCServices"><wsdl:port name="BasicHttpBinding_IIARCServices" binding="tns:BasicHttpBinding_IIARCServices"><soap:address location="https://www.globalratings.com/IARCPRODService/IARCServices.svc"/></wsdl:port></wsdl:service></wsdl:definitions>
