--- lib/iarc/wsdl/readme.txt
@@ -1,5 +1,0 @@
-A cached version of the WSDL. To update run:
-
-  ./manage.py refresh_wsdl
-
-To edit WSDL URLs, see mkt/developers/management/commands/refresh_wsdl.py
