--- lib/iarc_v2/__init__.py
@@ -1,1 +1,0 @@
-import client  # noqa
