+++ mkt/developers/templates/developers/apps/ratings/ratings_edit.html
@@ -64,131 +64,60 @@
             You can keep trying or try again later.
           {% endtrans %}
         </p>
-        {% if waffle.switch('iarc-upgrade-v2') %}
-          <form id="iarc" method="get" action="{{ settings.IARC_V2_SUBMISSION_ENDPOINT }}"
-                target="IARCForm">
-            <input type="hidden" name="StoreID" value="{{ settings.IARC_V2_STORE_ID }}">
-            <input type="hidden" name="StoreRequestID" value="{{ iarc_request_id }}">
-            <input type="hidden" name="LanguageID" value="{{ request.LANG }}">
-            <p class="listing-footer">
-              <button type="submit" class="create-iarc-rating button"
-                      data-api-url="{{ url('content-ratings-list', addon.app_slug)|urlparams(since=now) }}"
-                      data-redirect-url="{{ addon.get_dev_url('ratings') }}">
-                {{ _('Create an IARC Ratings Certificate') }}</button>
-            </p>
-          </form>
-        {% else %}
-          <form id="iarc" method="post" action="{{ settings.IARC_SUBMISSION_ENDPOINT }}" target="IARCForm">
-            <input type="hidden" name="email" value="{{ request.user.email }}">
-            <input type="hidden" name="company" value="{{ company }}">
-            <input type="hidden" name="appname" value="{{ app_name }}">
-            <input type="hidden" name="platform" value="{{ settings.IARC_PLATFORM }}">
-            <input type="hidden" name="storefront" value="{{ settings.IARC_STOREFRONT_ID }}">
-            <input type="hidden" name="token" value="{{ addon.iarc_token() }}">
-            <input type="hidden" name="pingbackurl" value="{{ url('content-ratings-pingback', addon.app_slug)|absolutify }}">
-            <p class="listing-footer">
-              <button type="submit" class="create-iarc-rating button"
-                      data-api-url="{{ url('content-ratings-list', addon.app_slug)|urlparams(since=now) }}"
-                      data-redirect-url="{{ addon.get_dev_url('ratings') }}">
-                {{ _('Create an IARC Ratings Certificate') }}</button>
-            </p>
-          </form>
-        {% endif %}
+        <form id="iarc" method="get" action="{{ settings.IARC_V2_SUBMISSION_ENDPOINT }}"
+              target="IARCForm">
+          <input type="hidden" name="StoreID" value="{{ settings.IARC_V2_STORE_ID }}">
+          <input type="hidden" name="StoreRequestID" value="{{ iarc_request_id }}">
+          <input type="hidden" name="LanguageID" value="{{ request.LANG }}">
+          <p class="listing-footer">
+            <button type="submit" class="create-iarc-rating button"
+                    data-api-url="{{ url('content-ratings-list', addon.app_slug)|urlparams(since=now) }}"
+                    data-redirect-url="{{ addon.get_dev_url('ratings') }}">
+              {{ _('Create an IARC Ratings Certificate') }}</button>
+          </p>
+        </form>
       </div>
     </section>
 
     <section>
       <h2>
-        {% if waffle.switch('iarc-upgrade-v2') %}
-          {{ _('Already Have an IARC Certificate?') }}
-        {% else %}
-          {{ _('Already Have Your App Rated?') }}
-        {% endif %}
+        {{ _('Already Have an IARC Certificate?') }}
       </h2>
 
       <div class="island">
-        {% if waffle.switch('iarc-upgrade-v2') %}
+        <p>
+          {% trans %}
+            If you have already gotten your app rated via the IARC Global
+            Ratings Tool above and already have an IARC certificate, you can
+            enter it here and we'll notify IARC to link your Marketplace
+            submission with that certificate.
+          {% endtrans %}
+        </p>
+        {% if addon.is_rated() %}
           <p>
             {% trans %}
-              If you have already gotten your app rated via the IARC Global
-              Ratings Tool above and already have an IARC certificate, you can
-              enter it here and we'll notify IARC to link your Marketplace
-              submission with that certificate.
-            {% endtrans %}
-          </p>
-          {% if addon.is_rated() %}
-            <p>
-              {% trans %}
-                Warning: if a certificate is already attached to your app, this
-                will overwrite it. Use with caution.
-              {% endtrans %}
-            </p>
-          {% endif %}
-        {% else %}
-          <p>
-            {% trans %}
-              If you have already gotten your app rated via the IARC Global
-              Ratings Tool above and already have an IARC security code, you
-              can manually enter it here.
+              Warning: if a certificate is already attached to your app, this
+              will overwrite it. Use with caution.
             {% endtrans %}
           </p>
         {% endif %}
         <form method="POST">
           {{ csrf() }}
-          {% if waffle.switch('iarc-upgrade-v2') %}
-            <div class="iarc-cert">
-              <label for="{{ form.cert_id.auto_id }}">
-                {{ _('Certificate ID') }}
-                {{ required() }}
-                {{ form.cert_id }}
-              </label>
-                  {{ tip(None,
-                         _("The certificate ID is an unique identifier sent "
-                           "to you. It is a mix of letters, numbers and "
-                           "dashes following the UUID canonical hexadecimal "
-                           "format like "
-                           "&quot;adb3261b-c657-4fd2-a057-bc9f85310b80&quot;."
-                          )) }}
-              {{ form.cert_id.errors }}
-            </div>
-          {% else %}
-            <table class="iarc-sec-code">
-              <tbody>
-                <tr>
-                  <th>
-                    <label for="{{ form.submission_id.auto_id }}">
-                      {{ _('Submission ID') }}
-                      {{ required() }}
-                    </label>
-                  </th>
-                  <td>
-                    {{ form.submission_id }}
-                    {{ tip(None,
-                           _("The submission ID can be found on the "
-                             "certificate sent to you. It is a numeric value "
-                             "like &quot;14&quot;.")) }}
-                    {{ form.submission_id.errors }}
-                  </td>
-                </tr>
-                <tr>
-                  <th>
-                    <label for="{{ form.security_code.auto_id }}">
-                      {{ _('Security Code') }}
-                      {{ required() }}
-                    </label>
-                  </th>
-                  <td>
-                    {{ form.security_code }}
-                    {{ tip(None,
-                           _("The security code can be found on the "
-                             "certificate sent to you. It is a mix of letters "
-                             "and numbers like &quot;AB12CD3&quot;.")) }}
-                    {{ form.security_code.errors }}
-                  </td>
-                </tr>
-              </tbody>
-            </table>
-          {% endif %}
+          <div class="iarc-cert">
+            <label for="{{ form.cert_id.auto_id }}">
+              {{ _('Certificate ID') }}
+              {{ required() }}
+              {{ form.cert_id }}
+            </label>
+                {{ tip(None,
+                       _("The certificate ID is an unique identifier sent "
+                         "to you. It is a mix of letters, numbers and "
+                         "dashes following the UUID canonical hexadecimal "
+                         "format like "
+                         "&quot;adb3261b-c657-4fd2-a057-bc9f85310b80&quot;."
+                        )) }}
+            {{ form.cert_id.errors }}
+          </div>
           {{ form.non_field_errors() }}
           <p class="listing-footer">
             <button>{{ _('Submit') }}</button>
