+++ mkt/developers/tests/test_commands.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 import mock
-from nose.tools import eq_, ok_
+from nose.tools import eq_
 
 import mkt
 import mkt.site.tests
+import mkt.site.utils
 from mkt.developers.management.commands import (cleanup_addon_premium,
                                                 exclude_unrated,
                                                 refresh_iarc_ratings)
 from mkt.site.fixtures import fixture
-from mkt.webapps.models import (AddonPremium, IARCCert,
-                                IARCInfo, RatingDescriptors, Webapp)
+from mkt.webapps.models import AddonPremium, IARCCert, Webapp
 
 
 class TestCommandViews(mkt.site.tests.TestCase):
@@ -49,7 +49,7 @@ class TestExcludeUnrated(mkt.site.tests.TestCase):
         assert not self._germany_listed()
 
     def test_dont_exclude_rated(self, index_mock):
-        mkt.site.tests.make_rated(self.webapp)
+        mkt.site.utils.make_rated(self.webapp)
 
         exclude_unrated.Command().handle()
         assert self._brazil_listed()
@@ -94,44 +94,7 @@ class TestRefreshIARCRatings(mkt.site.tests.TestCase):
     fixtures = fixture('webapp_337141')
 
     def setUp(self):
-        self.webapp = Webapp.objects.get(pk=337141)
-
-    def test_refresh_create(self):
-        IARCInfo.objects.create(
-            addon=self.webapp, submission_id=52, security_code='FZ32CU8')
-        refresh_iarc_ratings.Command().handle()
-
-        ok_(self.webapp.rating_descriptors)
-        ok_(self.webapp.rating_interactives)
-        ok_(self.webapp.content_ratings.count())
-
-    def test_refresh_update(self):
-        IARCInfo.objects.create(
-            addon=self.webapp, submission_id=52, security_code='FZ32CU8')
-        rd = RatingDescriptors.objects.create(
-            addon=self.webapp, has_usk_violence=True)
-        refresh_iarc_ratings.Command().handle()
-
-        ok_(rd.reload().has_esrb_strong_lang)
-        ok_(not rd.has_usk_violence)
-
-    def test_no_cert_no_refresh(self):
-        refresh_iarc_ratings.Command().handle()
-        ok_(not self.webapp.content_ratings.count())
-
-    def test_single_app(self):
-        IARCInfo.objects.create(
-            addon=self.webapp, submission_id=52, security_code='FZ32CU8')
-        refresh_iarc_ratings.Command().handle(apps=unicode(self.webapp.id))
-        ok_(self.webapp.content_ratings.count())
-
-
-class TestRefreshIARCRatingsV2(mkt.site.tests.TestCase):
-    fixtures = fixture('webapp_337141')
-
-    def setUp(self):
-        super(TestRefreshIARCRatingsV2, self).setUp()
-        self.create_switch('iarc-upgrade-v2')
+        super(TestRefreshIARCRatings, self).setUp()
         self.app = Webapp.objects.get(pk=337141)
 
     @mock.patch('mkt.developers.management.commands.refresh_iarc_ratings'
@@ -147,12 +110,5 @@ class TestRefreshIARCRatingsV2(mkt.site.tests.TestCase):
     @mock.patch('mkt.developers.management.commands.refresh_iarc_ratings'
                 '.refresh_iarc_ratings.delay')
     def test_refresh_does_not_have_cert(self, refresh_iarc_ratings_delay_mock):
-        refresh_iarc_ratings.Command().handle()
-        eq_(refresh_iarc_ratings_delay_mock.call_count, 0)
-
-        # Having IARCInfo objects should not matter, if we have no IARCCert,
-        # we can't refresh ratings.
-        IARCInfo.objects.create(
-            addon=self.app, submission_id=52, security_code='FZ32CU8')
         refresh_iarc_ratings.Command().handle()
         eq_(refresh_iarc_ratings_delay_mock.call_count, 0)
