+++ mkt/developers/urls.py
@@ -9,8 +9,7 @@ from lib.misc.urlconf_decorator import decorate
 from mkt.api.base import SubRouter
 from mkt.developers.api_payments import (PaymentCheckViewSet,
                                          PaymentDebugViewSet)
-from mkt.developers.views import (ContentRatingList, ContentRatingsPingback,
-                                  ContentRatingsPingbackV2)
+from mkt.developers.views import ContentRatingList, ContentRatingsPingback
 from mkt.inapp.views import InAppProductViewSet, StubInAppProductViewSet
 from mkt.receipts.urls import test_patterns
 from mkt.site.decorators import use_master
@@ -160,12 +159,8 @@ payments_api_patterns = patterns(
 
 dev_api_patterns = patterns(
     '',
-    url(r'^apps/app/(?P<pk>[^/<>"\']+)/content-ratings/pingback/',
-        ContentRatingsPingback.as_view(), name='content-ratings-pingback'),
-    # IARC v2 pushcert does not use the app pk, instead it uses IARCRequest
-    # uuid, but sends it as POST data.
-    url(r'^apps/iarc/pushcert/', ContentRatingsPingbackV2.as_view(),
-        name='content-ratings-pingback-v2'),
+    url(r'^apps/iarc/pushcert/', ContentRatingsPingback.as_view(),
+        name='content-ratings-pingback'),
     url(r'^apps/app/(?P<pk>[^/<>"\']+)/content-ratings/',
         ContentRatingList.as_view(), name='content-ratings-list'),
 )
