+++ mkt/developers/views.py
@@ -19,22 +19,18 @@ from django.views.decorators.http import require_POST
 
 import commonware.log
 import waffle
-from rest_framework import status as http_status
 from rest_framework.exceptions import ParseError
-from rest_framework.generics import CreateAPIView, GenericAPIView, ListAPIView
+from rest_framework.generics import GenericAPIView, ListAPIView
 from rest_framework.mixins import UpdateModelMixin
 from rest_framework.permissions import AllowAny
-from rest_framework.response import Response
 from rest_framework.status import is_success
 from session_csrf import anonymous_csrf, anonymous_csrf_exempt
 from django.utils.translation import ugettext as _
 from waffle.decorators import waffle_switch
 
 import mkt
-import lib.iarc
-from lib.iarc.utils import get_iarc_app_title
-from lib.iarc_v2.client import _iarc_app_data
-from lib.iarc_v2.serializers import IARCV2RatingListSerializer
+from lib.iarc.client import _iarc_app_data
+from lib.iarc.serializers import IARCRatingListSerializer
 from mkt.access import acl
 from mkt.api.base import CORSMixin, SlugOrIdMixin
 from mkt.api.models import Access
@@ -44,8 +40,7 @@ from mkt.developers.decorators import dev_required
 from mkt.developers.forms import (
     APIConsumerForm, AppFormBasic, AppFormDetails, AppFormMedia,
     AppFormSupport, AppFormTechnical, AppVersionForm, CategoryForm,
-    ContentRatingForm, IARCGetAppInfoForm, IARCV2ExistingCertificateForm,
-    MOTDForm,
+    ContentRatingForm, IARCExistingCertificateForm, MOTDForm,
     NewPackagedAppForm, PreviewFormSet, TransactionFilterForm, trap_duplicate)
 from mkt.developers.models import AppLog, IARCRequest
 from mkt.developers.serializers import ContentRatingSerializer
@@ -67,8 +62,7 @@ from mkt.users.models import UserProfile
 from mkt.users.views import _clean_next_url
 from mkt.versions.models import Version
 from mkt.webapps.decorators import app_view
-from mkt.webapps.models import (AddonUser, ContentRating, IARCInfo, IARCCert,
-                                Webapp)
+from mkt.webapps.models import AddonUser, ContentRating, IARCCert, Webapp
 from mkt.webapps.tasks import _update_manifest, update_manifests
 from mkt.zadmin.models import set_config, unmemoized_get_config
 
@@ -356,12 +350,11 @@ def content_ratings(request, addon_id, addon):
         'rating_interactives': addon.rating_interactives,
     }
 
-    if waffle.switch_is_active('iarc-upgrade-v2'):
-        # If there is a cert attached, we show it to the developer.
-        try:
-            ctx['cert'] = addon.iarc_cert
-        except IARCCert.DoesNotExist:
-            pass
+    # If there is a cert attached, we show it to the developer.
+    try:
+        ctx['cert'] = addon.iarc_cert
+    except IARCCert.DoesNotExist:
+        pass
 
     return render(request, 'developers/apps/ratings/ratings_summary.html', ctx)
 
@@ -375,24 +368,12 @@ def content_ratings_edit(request, addon_id, addon):
     data = request.POST if request.method == 'POST' else None
     template_name = 'developers/apps/ratings/ratings_edit.html'
 
-    if waffle.switch_is_active('iarc-upgrade-v2'):
-        form_class = IARCV2ExistingCertificateForm
-    else:
-        try:
-            app_info = addon.iarc_info
-            initial['submission_id'] = app_info.submission_id
-            initial['security_code'] = app_info.security_code
-        except IARCInfo.DoesNotExist:
-            pass
-        form_class = IARCGetAppInfoForm
-
     ctx = {}
 
-    form = form_class(data=data, initial=initial, app=addon)
+    form = IARCExistingCertificateForm(data=data, initial=initial, app=addon)
     if request.method == 'POST' and form.is_valid():
         try:
-            if (waffle.switch_is_active('iarc-upgrade-v2') and
-                    not form.cleaned_data.get('confirmed')):
+            if not form.cleaned_data.get('confirmed'):
                 # We need to show a confirmation to the user using the summary
                 # template. If a ValidationError is raised, it means the Cert
                 # ID was not recognized by IARC - we want to show the error in
@@ -418,27 +399,25 @@ def content_ratings_edit(request, addon_id, addon):
 
     ctx.update({
         'addon': addon,
-        'app_name': get_iarc_app_title(addon),
         'form': form,
         'company': addon.latest_version.developer_name,
         'now': datetime.now().strftime('%Y-%m-%d %H:%M:%S')
     })
 
-    if waffle.switch_is_active('iarc-upgrade-v2'):
-        try:
-            iarc_request = addon.iarc_request
-            outdated = (datetime.now() - iarc_request.created >
-                        timedelta(hours=1))
-            if outdated:
-                # IARC request outdated. Re-create.
-                iarc_request.delete()
-                iarc_request = IARCRequest.objects.create(
-                    app=addon, uuid=uuid.uuid4())
-        except IARCRequest.DoesNotExist:
-            # No IARC request exists. Create.
+    try:
+        iarc_request = addon.iarc_request
+        outdated = (datetime.now() - iarc_request.created >
+                    timedelta(hours=1))
+        if outdated:
+            # IARC request outdated. Re-create.
+            iarc_request.delete()
             iarc_request = IARCRequest.objects.create(
                 app=addon, uuid=uuid.uuid4())
-        ctx['iarc_request_id'] = unicode(uuid.UUID(iarc_request.uuid))
+    except IARCRequest.DoesNotExist:
+        # No IARC request exists. Create.
+        iarc_request = IARCRequest.objects.create(
+            app=addon, uuid=uuid.uuid4())
+    ctx['iarc_request_id'] = unicode(uuid.UUID(iarc_request.uuid))
 
     return render(request, template_name, ctx)
 
@@ -1118,105 +1097,8 @@ class ContentRatingList(CORSMixin, SlugOrIdMixin, ListAPIView):
         return super(ContentRatingList, self).get(self, request)
 
 
-class ContentRatingsPingback(CORSMixin, SlugOrIdMixin, CreateAPIView):
-    cors_allowed_methods = ['post']
-    parser_classes = (lib.iarc.utils.IARC_JSON_Parser,)
-    permission_classes = (AllowAny,)
-
-    queryset = Webapp.objects.all()
-    slug_field = 'app_slug'
-
-    def post(self, request, pk, *args, **kwargs):
-        log.info(u'Received IARC pingback for app:%s' % pk)
-
-        if request.content_type != 'application/json':
-            log.info(u'IARC pingback not of content-type "application/json"')
-            return Response({
-                'detail': "Endpoint only accepts 'application/json'."
-            }, status=http_status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)
-
-        app = self.get_object()
-        data = request.data[0]
-        if settings.DEBUG:
-            log.debug(u'%s' % data)
-
-        if app.iarc_token() != data.get('token'):
-            # Verify token.
-            log.info(u'Token mismatch in IARC pingback for app:%s' % app.id)
-            return Response({'detail': 'Token mismatch'},
-                            status=http_status.HTTP_400_BAD_REQUEST)
-
-        if data.get('ratings'):
-            # Double-check with IARC that it's the correct rating.
-            if not self.verify_data(data):
-                return Response('The ratings do not match the submission ID.',
-                                status=http_status.HTTP_400_BAD_REQUEST)
-
-            log.info(u'Setting content ratings from IARC pingback for app:%s' %
-                     app.id)
-            # We found a rating, so store the id and code for future use.
-            if 'submission_id' in data and 'security_code' in data:
-                app.set_iarc_info(data['submission_id'], data['security_code'])
-
-            # Update status if incomplete status.
-            # Do this before set_content_ratings to not prematurely trigger
-            # a refresh.
-            log.info('Checking app:%s completeness after IARC pingback.'
-                     % app.id)
-            if (app.has_incomplete_status() and
-                    app.is_fully_complete(ignore_ratings=True)):
-                log.info('Updating app status from IARC pingback for app:%s' %
-                         app.id)
-                # Don't call update to prevent recursion in update_status.
-                app.update(status=mkt.STATUS_PENDING)
-                log.info('Updated app status from IARC pingback for app:%s' %
-                         app.id)
-            elif app.has_incomplete_status():
-                log.info('Reasons for app:%s incompleteness after IARC '
-                         'pingback: %s' % (app.id, app.completion_errors()))
-
-            app.set_descriptors(data.get('descriptors', []))
-            app.set_interactives(data.get('interactives', []))
-            # Set content ratings last since it triggers a refresh on Content
-            # Ratings page. We want descriptors and interactives visible by
-            # the time it's refreshed.
-            app.set_content_ratings(data.get('ratings', {}))
-
-        return Response('ok')
-
-    def verify_data(self, data):
-        client = lib.iarc.client.get_iarc_client('services')
-        xml = lib.iarc.utils.render_xml('get_app_info.xml', data)
-        resp = client.Get_App_Info(XMLString=xml)
-        check_data = lib.iarc.utils.IARC_XML_Parser().parse_string(resp)
-        try:
-            check_data = check_data.get('rows', [])[0]
-        except IndexError:
-            return False
-
-        rates_bad = data.get('ratings') != check_data.get('ratings')
-        inter_bad = (set(data.get('interactives', [])) !=
-                     set(check_data.get('interactives', [])))
-        descs_bad = (set(data.get('descriptors', [])) !=
-                     set(check_data.get('descriptors', [])))
-        if rates_bad:
-            log.error('IARC pingback did not match rating %s vs %s' %
-                      (data.get('ratings'), check_data.get('ratings')))
-        if inter_bad:
-            log.error('IARC pingback did not match interactives %s vs %s' %
-                      (data.get('interactives'),
-                       check_data.get('interactives')))
-        if descs_bad:
-            log.error('IARC pingback did not match descriptors %s vs %s' %
-                      (data.get('descriptors'), check_data.get('descriptors')))
-        if rates_bad or inter_bad or descs_bad:
-            return False
-
-        return True
-
-
-class ContentRatingsPingbackV2(CORSMixin, UpdateModelMixin, GenericAPIView):
-    """Pingback API for IARC v2.
+class ContentRatingsPingback(CORSMixin, UpdateModelMixin, GenericAPIView):
+    """Pingback API for IARC.
 
     Should conform to the PushCert API spec. Assumes that the RatingList is
     always sent with PushCert, so we don't need to call SearchCert afterwards
@@ -1224,7 +1106,7 @@ class ContentRatingsPingbackV2(CORSMixin, UpdateModelMixin, GenericAPIView):
     cors_allowed_methods = ['post']
     permission_classes = (AllowAny,)
     queryset = IARCRequest.objects.all()
-    serializer_class = IARCV2RatingListSerializer
+    serializer_class = IARCRatingListSerializer
     lookup_field = 'uuid'
 
     def get_object(self, queryset=None):
@@ -1236,7 +1118,7 @@ class ContentRatingsPingbackV2(CORSMixin, UpdateModelMixin, GenericAPIView):
             raise ParseError('Need a StoreRequestID')
         except ValueError:
             raise ParseError('StoreRequestID is not a valid UUID')
-        self.object = super(ContentRatingsPingbackV2, self).get_object().app
+        self.object = super(ContentRatingsPingback, self).get_object().app
         return self.object
 
     def finalize_response(self, request, response, *args, **kwargs):
@@ -1253,7 +1135,7 @@ class ContentRatingsPingbackV2(CORSMixin, UpdateModelMixin, GenericAPIView):
                 pass
         else:
             response.data['StatusCode'] = 'InvalidRequest'
-        return super(ContentRatingsPingbackV2, self).finalize_response(
+        return super(ContentRatingsPingback, self).finalize_response(
             request, response, *args, **kwargs)
 
     def post(self, request, *args, **kwargs):
