+++ mkt/reviewers/tests/test_views.py
@@ -49,7 +49,7 @@ from mkt.site.helpers import absolutify, isotime
 from mkt.site.storage_utils import private_storage, public_storage
 from mkt.site.tests import (check_links, days_ago, formset, initial,
                             req_factory_factory, user_factory)
-from mkt.site.utils import app_factory, make_game, paginate, version_factory
+from mkt.site.utils import app_factory, make_rated, paginate, version_factory
 from mkt.submit.tests.test_views import BasePackagedAppTest, SetupFilesMixin
 from mkt.tags.models import Tag
 from mkt.users.models import UserProfile
@@ -1359,7 +1359,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         super(TestReviewApp, self).setUp()
         self.mozilla_contact = 'contact@mozilla.com'
         self.app = self.get_app()
-        self.app = make_game(self.app, True)
+        make_rated(self.app)
         self.app.update(status=mkt.STATUS_PENDING,
                         mozilla_contact=self.mozilla_contact)
         self.version = self.app.latest_version
@@ -1421,6 +1421,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         response = self.client.get(self.url)
         eq_(response.status_code, 200)
 
+    @mock.patch('mkt.reviewers.utils.iarc_publish.delay', lambda app_id: None)
     def test_sr_can_review_blocklisted_app(self):
         self.app.update(status=mkt.STATUS_BLOCKED)
         self.login_as_senior_reviewer()
@@ -1452,8 +1453,8 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         self._check_email_dev_and_contact('Rejected')
         self._check_email_body()
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
-    def test_pending_to_public_w_requirements_overrides(self, storefront_mock):
+    @mock.patch('mkt.reviewers.utils.iarc_publish')
+    def test_pending_to_public_w_requirements_overrides(self, publish_mock):
         data = {'action': 'public', 'comments': 'something',
                 'has_packaged_apps': True}
         data.update(self._attachment_management_form(num=0))
@@ -1471,12 +1472,12 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         # A reviewer changing features shouldn't generate a re-review.
         eq_(RereviewQueue.objects.count(), 0)
 
-        # Since the app is not public yet (see above), there is no need to
-        # notify IARC.
-        eq_(storefront_mock.call_count, 0)
+        # Since the app is not really public yet (see above), there is no need
+        # to notify IARC.
+        eq_(publish_mock.delay.call_count, 0)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
-    def test_pending_to_public_w_requirements_removed(self, storefront_mock):
+    @mock.patch('mkt.reviewers.utils.iarc_publish')
+    def test_pending_to_public_w_requirements_removed(self, publish_mock):
         self.app.latest_version.features.update(has_packaged_apps=True)
         data = {'action': 'public', 'comments': 'something',
                 'has_packaged_apps': False}
@@ -1495,9 +1496,9 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         # A reviewer changing features shouldn't generate a re-review.
         eq_(RereviewQueue.objects.count(), 0)
 
-        # Since the app is not public yet (see above), there is no need to
-        # notify IARC.
-        eq_(storefront_mock.call_count, 0)
+        # Since the app is not really public yet (see above), there is no need
+        # to notify IARC.
+        eq_(publish_mock.delay.call_count, 0)
 
     def test_pending_to_reject_w_requirements_overrides(self):
         # Rejecting an app doesn't let you override features requirements.
@@ -1512,7 +1513,8 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         eq_(app.publish_type, mkt.PUBLISH_IMMEDIATE)
         eq_(app.status, mkt.STATUS_REJECTED)
 
-    def test_pending_to_reject_w_requirements_overrides_nothing_changed(self):
+    @mock.patch('mkt.reviewers.utils.iarc_publish.delay', lambda app_id: None)
+    def test_pending_to_public_w_requirements_overrides_nothing_changed(self):
         self.version.features.update(has_packaged_apps=True)
         data = {'action': 'public', 'comments': 'something',
                 'has_packaged_apps': True}
@@ -1597,14 +1599,8 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         self._check_email_body()
         self._check_score(mkt.REVIEWED_WEBAPP_HOSTED)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
-    def test_pending_to_public_iarc_v1(self, set_iarc_storefront_data_mock):
-        self._test_pending_to_public()
-        eq_(set_iarc_storefront_data_mock.call_count, 1)
-
     @mock.patch('mkt.reviewers.utils.iarc_publish')
-    def test_pending_to_public_iarc_v2(self, iarc_publish_mock):
-        self.create_switch('iarc-upgrade-v2')
+    def test_pending_to_public_iarc(self, iarc_publish_mock):
         self._test_pending_to_public()
         eq_(iarc_publish_mock.delay.call_count, 1)
         eq_(iarc_publish_mock.delay.call_args[0], (self.app.pk, ))
@@ -1627,6 +1623,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
 
         eq_(messages.call_args_list[0][0][1], 'Review successfully processed.')
 
+    @mock.patch('mkt.reviewers.utils.iarc_unpublish.delay', lambda appid: None)
     def test_pending_to_disable_senior_reviewer(self):
         self.login_as_senior_reviewer()
 
@@ -1671,14 +1668,8 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         self._check_email_dev_and_contact('Approved')
         self._check_email_body()
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
-    def test_escalation_to_public_iarc_v1(self, set_iarc_storefront_data_mock):
-        self._test_escalation_to_public()
-        eq_(set_iarc_storefront_data_mock.call_count, 1)
-
     @mock.patch('mkt.reviewers.utils.iarc_publish')
-    def test_escalation_to_public_iarc_v2(self, iarc_publish_mock):
-        self.create_switch('iarc-upgrade-v2')
+    def test_escalation_to_public_iarc(self, iarc_publish_mock):
         self._test_escalation_to_public()
         eq_(iarc_publish_mock.delay.call_count, 1)
         eq_(iarc_publish_mock.delay.call_args[0], (self.app.pk, ))
@@ -1701,6 +1692,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         self._check_email_body()
         self._check_score(mkt.REVIEWED_WEBAPP_HOSTED)
 
+    @mock.patch('mkt.reviewers.utils.iarc_unpublish.delay', lambda appid: None)
     def test_escalation_to_disable_senior_reviewer(self):
         self.login_as_senior_reviewer()
         EscalationQueue.objects.create(addon=self.app)
@@ -1761,6 +1753,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         self._check_email_body()
         self._check_score(mkt.REVIEWED_WEBAPP_REREVIEW)
 
+    @mock.patch('mkt.reviewers.utils.iarc_unpublish.delay', lambda appid: None)
     def test_rereview_to_disable_senior_reviewer(self):
         self.login_as_senior_reviewer()
 
@@ -2031,6 +2024,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         return data
 
     @override_settings(REVIEWER_ATTACHMENTS_PATH=ATTACHMENTS_DIR)
+    @mock.patch('mkt.reviewers.utils.iarc_publish.delay', lambda app_id: None)
     @mock.patch('mkt.site.storage_utils.LocalFileStorage.save')
     def test_no_attachments(self, save_mock):
         """ Test addition of no attachment """
@@ -2055,6 +2049,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         assert '<script>alert(42)</script>' not in response.content
         assert '&lt;script&gt;alert(42)&lt;/script&gt;' in response.content
 
+    @mock.patch('mkt.reviewers.utils.iarc_publish.delay', lambda app_id: None)
     def test_priority_flag_cleared_for_public(self):
         self.get_app().update(priority_review=True)
         data = {'action': 'public', 'comments': 'something'}
@@ -2063,6 +2058,7 @@ class TestReviewApp(SetupFilesMixin, AppReviewerTest, TestReviewMixin,
         self.post(data)
         eq_(self.get_app().priority_review, False)
 
+    @mock.patch('mkt.reviewers.utils.iarc_unpublish.delay', lambda appid: None)
     def test_priority_flag_uncleared_for_reject(self):
         self.get_app().update(priority_review=True)
         data = {'action': 'reject', 'comments': 'something'}
@@ -2159,7 +2155,7 @@ class TestCannedResponses(AppReviewerTest):
         assert self.cr.response in choices[0]
 
 
-@mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
+@mock.patch('mkt.reviewers.utils.iarc_publish')
 @mock.patch('mkt.reviewers.views.messages.success')
 @mock.patch('mkt.webapps.tasks.index_webapps')
 @mock.patch('mkt.webapps.tasks.update_cached_manifests')
@@ -2196,12 +2192,12 @@ class TestApproveHostedApp(AppReviewerTest, TestReviewMixin,
 
     def test_pending_to_public(self, update_name, update_locales,
                                update_cached_manifests,
-                               index_webapps, messages, storefront_mock):
+                               index_webapps, messages, publish_mock):
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2222,17 +2218,17 @@ class TestApproveHostedApp(AppReviewerTest, TestReviewMixin,
         eq_(index_webapps.delay.call_count, 1)
         # App is not packaged, no need to call update_cached_manifests.
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 1)
+        eq_(publish_mock.delay.call_count, 1)
 
     def test_pending_to_hidden(self, update_name, update_locales,
                                update_cached_manifests, index_webapps,
-                               messages, storefront_mock):
+                               messages, publish_mock):
         self.get_app().update(publish_type=mkt.PUBLISH_HIDDEN)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2253,11 +2249,11 @@ class TestApproveHostedApp(AppReviewerTest, TestReviewMixin,
         eq_(index_webapps.delay.call_count, 1)
         # App is not packaged, no need to call update_cached_manifests.
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 1)
+        eq_(publish_mock.delay.call_count, 1)
 
     def test_pending_to_approved(self, update_name, update_locales,
                                  update_cached_manifests, index_webapps,
-                                 messages, storefront_mock):
+                                 messages, publish_mock):
         self.get_app().update(publish_type=mkt.PUBLISH_PRIVATE)
         index_webapps.reset_mock()
         eq_(update_name.call_count, 0)
@@ -2287,12 +2283,12 @@ class TestApproveHostedApp(AppReviewerTest, TestReviewMixin,
         # App is not packaged, no need to call update_cached_manifests.
         eq_(update_cached_manifests.delay.call_count, 0)
         # App is private so we don't send this yet.
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(index_webapps.delay.call_count, 1)
 
     def test_pending_to_reject(self, update_name, update_locales,
                                update_cached_manifests, index_webapps,
-                               messages, storefront_mock):
+                               messages, publish_mock):
         index_webapps.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
@@ -2317,12 +2313,12 @@ class TestApproveHostedApp(AppReviewerTest, TestReviewMixin,
         eq_(update_name.call_count, 0)  # Not a packaged app.
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(index_webapps.delay.call_count, 1)
 
 
 @mock.patch('lib.crypto.packaged.sign')
-@mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
+@mock.patch('mkt.reviewers.utils.iarc_publish')
 @mock.patch('mkt.reviewers.views.messages.success')
 @mock.patch('mkt.webapps.tasks.index_webapps')
 @mock.patch('mkt.webapps.tasks.update_cached_manifests')
@@ -2360,12 +2356,12 @@ class TestApprovePackagedApp(AppReviewerTest, TestReviewMixin,
 
     def test_pending_to_public(self, update_name, update_locales,
                                update_cached_manifests, index_webapps,
-                               messages, storefront_mock, sign_mock):
+                               messages, publish_mock, sign_mock):
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2385,18 +2381,18 @@ class TestApprovePackagedApp(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 1)
+        eq_(publish_mock.delay.call_count, 1)
         eq_(sign_mock.call_args[0][0], self.get_app().current_version.pk)
 
     def test_pending_to_hidden(self, update_name, update_locales,
                                update_cached_manifests, index_webapps,
-                               messages, storefront_mock, sign_mock):
+                               messages, publish_mock, sign_mock):
         self.get_app().update(publish_type=mkt.PUBLISH_HIDDEN)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2416,18 +2412,18 @@ class TestApprovePackagedApp(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 1)
+        eq_(publish_mock.delay.call_count, 1)
         eq_(sign_mock.call_args[0][0], self.get_app().current_version.pk)
 
     def test_pending_to_approved(self, update_name, update_locales,
                                  update_cached_manifests, index_webapps,
-                                 messages, storefront_mock, sign_mock):
+                                 messages, publish_mock, sign_mock):
         self.get_app().update(publish_type=mkt.PUBLISH_PRIVATE)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2447,17 +2443,17 @@ class TestApprovePackagedApp(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_args[0][0], self.get_app().current_version.pk)
 
     def test_pending_to_rejected(self, update_name, update_locales,
                                  update_cached_manifests, index_webapps,
-                                 messages, storefront_mock, sign_mock):
+                                 messages, publish_mock, sign_mock):
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'reject', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2476,12 +2472,12 @@ class TestApprovePackagedApp(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 0)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_count, 0)
 
     def test_pending_to_approved_app_private_prior_version_rejected(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         """
         Test that everything works out ok when v1.0 was rejected and developer
         submitted v1.1 that is then approved. This should still be considered a
@@ -2499,7 +2495,7 @@ class TestApprovePackagedApp(AppReviewerTest, TestReviewMixin,
         update_cached_manifests.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         eq_(self.app.current_version, None)
         eq_(self.app.latest_version, self.new_version)
@@ -2524,12 +2520,12 @@ class TestApprovePackagedApp(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_args[0][0], self.new_version.pk)
 
 
 @mock.patch('lib.crypto.packaged.sign')
-@mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
+@mock.patch('mkt.reviewers.utils.iarc_publish')
 @mock.patch('mkt.reviewers.views.messages.success')
 @mock.patch('mkt.webapps.tasks.index_webapps')
 @mock.patch('mkt.webapps.tasks.update_cached_manifests')
@@ -2570,12 +2566,12 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
 
     def test_version_pending_to_public(self, update_name, update_locales,
                                        update_cached_manifests, index_webapps,
-                                       messages, storefront_mock, sign_mock):
+                                       messages, publish_mock, sign_mock):
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2596,19 +2592,19 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 1)
+        eq_(publish_mock.delay.call_count, 1)
         eq_(sign_mock.call_args[0][0], app.current_version.pk)
 
     def test_version_pending_to_approved(self, update_name, update_locales,
                                          update_cached_manifests,
                                          index_webapps, messages,
-                                         storefront_mock, sign_mock):
+                                         publish_mock, sign_mock):
         self.app.update(publish_type=mkt.PUBLISH_PRIVATE)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2630,18 +2626,18 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_args[0][0], self.new_version.pk)
 
     def test_version_pending_to_public_app_unlisted(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         self.app.update(status=mkt.STATUS_UNLISTED)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2662,19 +2658,19 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 1)
+        eq_(publish_mock.delay.call_count, 1)
         eq_(sign_mock.call_args[0][0], app.current_version.pk)
 
     def test_version_pending_to_approved_app_unlisted(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         self.app.update(status=mkt.STATUS_UNLISTED,
                         publish_type=mkt.PUBLISH_PRIVATE)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2696,18 +2692,18 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_args[0][0], self.new_version.pk)
 
     def test_version_pending_to_public_app_private(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         self.app.update(status=mkt.STATUS_APPROVED)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2728,19 +2724,19 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 1)
+        eq_(publish_mock.delay.call_count, 1)
         eq_(sign_mock.call_args[0][0], app.current_version.pk)
 
     def test_version_pending_to_approved_app_private(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         self.app.update(status=mkt.STATUS_APPROVED,
                         publish_type=mkt.PUBLISH_PRIVATE)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'public', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2762,18 +2758,18 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 1)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 1)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_args[0][0], self.new_version.pk)
 
     def test_version_pending_to_rejected_app_public(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         self.app.update(status=mkt.STATUS_PUBLIC)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'reject', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2795,18 +2791,18 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 0)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_count, 0)
 
     def test_version_pending_to_rejected_app_unlisted(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         self.app.update(status=mkt.STATUS_UNLISTED)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'reject', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2828,18 +2824,18 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 0)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_count, 0)
 
     def test_version_pending_to_rejected_app_private(
             self, update_name, update_locales, update_cached_manifests,
-            index_webapps, messages, storefront_mock, sign_mock):
+            index_webapps, messages, publish_mock, sign_mock):
         self.app.update(status=mkt.STATUS_APPROVED)
         index_webapps.delay.reset_mock()
         eq_(update_name.call_count, 0)
         eq_(update_locales.call_count, 0)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
 
         data = {'action': 'reject', 'comments': 'something'}
         data.update(self._attachment_management_form(num=0))
@@ -2861,7 +2857,7 @@ class TestApprovePackagedVersions(AppReviewerTest, TestReviewMixin,
         eq_(update_locales.call_count, 0)
         eq_(index_webapps.delay.call_count, 1)
         eq_(update_cached_manifests.delay.call_count, 0)
-        eq_(storefront_mock.call_count, 0)
+        eq_(publish_mock.delay.call_count, 0)
         eq_(sign_mock.call_count, 0)
 
 
@@ -3116,6 +3112,8 @@ class TestMotd(AppReviewerTest, AccessMixin):
         eq_(get_config(self.key), u'new motd')
 
 
+@mock.patch('mkt.reviewers.utils.iarc_publish.delay', lambda app_id: None)
+@mock.patch('mkt.reviewers.utils.iarc_unpublish.delay', lambda app_id: None)
 class TestReviewAppComm(AppReviewerTest, AttachmentManagementMixin,
                         TestReviewMixin, TestedonManagementMixin):
     """
