+++ mkt/webapps/tests/test_tasks.py
@@ -315,8 +315,7 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
         assert not retry.called
         assert RereviewQueue.objects.filter(addon=self.addon).exists()
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
-    def test_manifest_validation_failure(self, _iarc):
+    def test_manifest_validation_failure(self):
         # We are already mocking validator, but this test needs to make sure
         # it actually saves our custom validation result, so add that.
         def side_effect(upload_id, **kwargs):
@@ -355,11 +354,9 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
         ok_(msg.subject.startswith('Issue with your app'))
         ok_(validation_results['messages'][0]['message'] in msg.body)
         ok_(validation_url in msg.body)
-        ok_(not _iarc.called)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_manifest_name_change_rereview(self, _manifest, _iarc):
+    def test_manifest_name_change_rereview(self, _manifest):
         # Mock original manifest file lookup.
         _manifest.return_value = original
         # Mock new manifest with name change.
@@ -370,11 +367,9 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
         eq_(RereviewQueue.objects.count(), 1)
         # 2 logs: 1 for manifest update, 1 for re-review trigger.
         eq_(ActivityLog.objects.for_apps([self.addon]).count(), 2)
-        ok_(_iarc.called)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_manifest_locale_name_add_rereview(self, _manifest, _iarc):
+    def test_manifest_locale_name_add_rereview(self, _manifest):
         # Mock original manifest file lookup.
         _manifest.return_value = original
         # Mock new manifest with name change.
@@ -389,11 +384,9 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
             action=mkt.LOG.REREVIEW_MANIFEST_CHANGE.id)[0]
         eq_(log.details.get('comments'),
             u'Locales added: "eso" (es).')
-        ok_(not _iarc.called)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_manifest_locale_name_change_rereview(self, _manifest, _iarc):
+    def test_manifest_locale_name_change_rereview(self, _manifest):
         # Mock original manifest file lookup.
         _manifest.return_value = original
         # Mock new manifest with name change.
@@ -408,11 +401,9 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
             action=mkt.LOG.REREVIEW_MANIFEST_CHANGE.id)[0]
         eq_(log.details.get('comments'),
             u'Locales updated: "Mozilla Kugel" -> "Bippity Bop" (de).')
-        ok_(not _iarc.called)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_manifest_default_locale_change(self, _manifest, _iarc):
+    def test_manifest_default_locale_change(self, _manifest):
         # Mock original manifest file lookup.
         _manifest.return_value = original
         # Mock new manifest with name change.
@@ -432,11 +423,9 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
             u'Manifest name changed from "MozillaBall" to "Mozilla Balón". '
             u'Default locale changed from "en-US" to "es". '
             u'Locales added: "Mozilla Balón" (es).')
-        ok_(_iarc.called)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_manifest_locale_name_removal_no_rereview(self, _manifest, _iarc):
+    def test_manifest_locale_name_removal_no_rereview(self, _manifest):
         # Mock original manifest file lookup.
         _manifest.return_value = original
         # Mock new manifest with name change.
@@ -450,11 +439,9 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
         eq_(RereviewQueue.objects.count(), 0)
         # Log for manifest update.
         eq_(ActivityLog.objects.for_apps([self.addon]).count(), 1)
-        ok_(not _iarc.called)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_force_rereview(self, _manifest, _iarc):
+    def test_force_rereview(self, _manifest):
         # Mock original manifest file lookup.
         _manifest.return_value = original
         # Mock new manifest with name change.
@@ -472,11 +459,8 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
         # 2 logs: 1 for manifest update, 1 for re-review trigger.
         eq_(ActivityLog.objects.for_apps([self.addon]).count(), 2)
 
-        ok_(_iarc.called)
-
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_manifest_support_locales_change(self, _manifest, _iarc):
+    def test_manifest_support_locales_change(self, _manifest):
         """
         Test both PUBLIC and PENDING to catch apps w/o `current_version`.
         """
@@ -491,11 +475,9 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
             self._run()
             ver = self.version.reload()
             eq_(ver.supported_locales, 'de,es,fr')
-            ok_(not _iarc.called)
 
-    @mock.patch('mkt.webapps.models.Webapp.set_iarc_storefront_data')
     @mock.patch('mkt.webapps.models.Webapp.get_manifest_json')
-    def test_manifest_support_developer_change(self, _manifest, _iarc):
+    def test_manifest_support_developer_change(self, _manifest):
         # Mock original manifest file lookup.
         _manifest.return_value = original
         # Mock new manifest with developer name change.
@@ -509,8 +491,6 @@ class TestUpdateManifest(mkt.site.tests.TestCase):
         eq_(RereviewQueue.objects.count(), 1)
         # 2 logs: 1 for manifest update, 1 for re-review trigger.
         eq_(ActivityLog.objects.for_apps([self.addon]).count(), 2)
-
-        ok_(_iarc.called)
 
 
 class TestDumpApps(mkt.site.tests.TestCase):
